change runNote from event to function call
When using events, we can't guarantee that the paragraphs will be executed in order, so we need to loop through the paragraphs and make them run.

The back-end is already taking care of execution of query in parallel or in order of call, this allows us to send the front-end calls to backend in order.

Author: Damien Corneau <corneadoug@gmail.com>

Closes #171 from corneadoug/fix/runAllParagraphsOrder and squashes the following commits:

6b5ecc8 [Damien Corneau] change runNote from event to function call