fix remoteinterpreterserver deadlock problem
After using zeppelin server for some days, we often found that paragraph can't run while the zeppelin server process is on, restart interpreter is just not work. Then we find out that the RemoteInterpreterServer is in deadlock, and can't be stopped by zeppelin server. Unfortunately, this deadlocked RemoteInterpreterServer will still hold the hadoop resources, we need to kill the process manually each time. The detail Info shows in: https://issues.apache.org/jira/browse/ZEPPELIN-271
We figure out that the deadlock is caused by the lack of synchronized (intpGroup) in SparkSqlInterpreter.java

Author: 曾林西 <zenglinxi@meituan.com>

Closes #267 from zenglinxi0615/fix-remoteinterpreterserver-deadlock and squashes the following commits:

a173374 [曾林西] fix a problem in getSparkInterpreter()
160fa02 [曾林西] move ((LazyOpenInterpreter) p).open() outside of synchronized(intpGroup) block to fix the remoteinterpreterserver deadlock
93e703d [曾林西] add synchronized (intpGroup) in open() of LazyOpenInterpreter.java, to make sure thread-locking sequence consistent between multithreading
4ec953b [曾林西] fix remoteinterpreterserver deadlock problem