Fix NotebookList Ordering
In the notebook list, the notebooks that have been created without name (empty name) and prefixed with 'Note' are not ordered correctly with the others.

The reason behind it, is that when notebook doesn't have a name, we use the notebook id and prefix it by 'Note'. At that time the list was already ordered and therefore the Name shown doesn't match the position in the list.

Before:
![screen shot 2015-08-22 at 2 48 56 pm](https://cloud.githubusercontent.com/assets/710411/9422740/f9d7a496-48dc-11e5-9326-bfddd16dd66d.png)

After:
![screen shot 2015-08-22 at 2 30 26 pm](https://cloud.githubusercontent.com/assets/710411/9422688/6fdc69f4-48da-11e5-8f39-5752369942ab.png)

The Fix I have is to have an ordering function at the place we prefix the 'Note' on the front side.
Another solution could be to just to have by default the notebook name as 'Note ' + note.id when we create a note. The second option could probably be better performance wise, however it wouldn't be backward compatible.

Author: Damien Corneau <corneadoug@gmail.com>

Closes #245 from corneadoug/fix/NotebookListOrder and squashes the following commits:

024bacb [Damien Corneau] Add getId function
e8a5e58 [Damien Corneau] Set name to Note + note.id on note creation
d585a0d [Damien Corneau] Add ordering service and order notebook list