Generate unique temporary files for Python and R Spark Interpreters.
### What is this PR for?
Generate unique temporary files for Python and R Spark interpreters using java.io.File.createTempFile.
Remove the temporary file for the Python interpreter when the interpreter is closed (is already being done in the R interpreter).
Without this fix it is not possible to run more than one Zeppelin interpreter on the same server due to file owner conflicts between instances.

### What type of PR is it?
Bug Fix

### Todos

### What is the Jira issue?
No Jira opened for this

### How should this be tested?

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? no
* Is there breaking changes for older versions? no
* Does this needs documentation? no

Author: Johan Muedsam <jm837u@att.com>

Closes #823 from mudsam/master and squashes the following commits:

0c018b4 [Johan Muedsam] Generate unique temporary files for Python and R Spark Interpreters. Add cleanup of Python temporary file on close().