Fix a few image not rendering
### What is this PR for?
There was a recent PR made by doanduyhai that was changing some images URL to include `../assets`. However, I found a few more cases in the deployed website.
It is actually a bit hard to understand why this happens in some cases only, maybe it could be better to replace all `/assets` to `../assets`, however this PR is only fixing the current issues.

### What type of PR is it?
Bug Fix

### How should this be tested?
You can build the doc and run it locally

### Screenshots (if appropriate)

<img width="688" alt="screen shot 2016-05-18 at 3 20 53 pm" src="https://cloud.githubusercontent.com/assets/710411/15376817/45648b78-1d0c-11e6-82cb-4ba106e01043.png">
<img width="521" alt="screen shot 2016-05-18 at 3 21 09 pm" src="https://cloud.githubusercontent.com/assets/710411/15376819/493ff976-1d0c-11e6-9960-cbd568f5d0ec.png">

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #899 from corneadoug/fix/docsImagesLink and squashes the following commits:

c7333d7 [Damien CORNEAU] Fix a few image not rendering