[ZEPPELIN-975] Move download data button outside the paragraph settings
### What is this PR for?
Following the PR #714 and its comments, we are moving the download data outside the paragraph settings. Which would allow a possible drop down to choose backend data download in the future.

### What type of PR is it?
Improvement

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-975

### How should this be tested?
Show a %Table type of result (you can use below code in a paragraph for that).
You should see the icon, click on it to download the data.
```
print(s"""%table
name\tsize
sun\t127.6
moon\t0.6458
gerard\t45.234856803""")
```

### Screenshots (if appropriate)
![screen shot 2016-06-08 at 4 15 22 pm](https://cloud.githubusercontent.com/assets/710411/15886052/bf3f8648-2d95-11e6-837f-47f112a76031.png)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No, but maybe feature announcement during release

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #975 from corneadoug/improve/downloadDataBtnVisibility and squashes the following commits:

2ca3028 [Damien CORNEAU] Move download data button outside the paragraph settings