[ZEPPELIN-1183] Replace $window.location by $location
### What is this PR for?
There is a native service used across the app to take care of re-routing to different page: `$location`, however there has been a few times where `$window.location` is used instead, and people tend to do it in new PR too.

This convert the `$window.location` from the code to use `$location`.

Only one was kept in `app.controller.js` for Iframe, so it doesn't flicker.

### What type of PR is it?
Refactoring

### What is the Jira issue?
https://issues.apache.org/jira/browse/ZEPPELIN-1183

### How should this be tested?
You can try to reproduce the couple of cases that redirection is needed,
but overall there shouldn't be any issue from that change.

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? No

Author: Damien CORNEAU <corneadoug@gmail.com>

Closes #1188 from corneadoug/ZEPPELIN-1183 and squashes the following commits:

1bc5f3b [Damien CORNEAU] unify the  paths
469f221 [Damien CORNEAU] Remove unused controller function parameters
60cd327 [Damien CORNEAU] Remove the window.location from the code, for