[ZEPPELIN-1294] Implement one-way sync for notebook repos
### What is this PR for?
Currently, if there are two notebook storages configured, the sync process treats them equally and can pull changes from either storage. This can cause confusions if there is a clear distinction between primary and secondary storages, and the primary storage is managed outside of Zeppelin itself, such as a local folder that's git controlled. If a notebook is deleted or reverted to an older version in the primary storage behind Zeppelin's back, the sync process will overwrite the change with data from the secondary storage, likely against the user's intention.

The proposal is to enable the scenario with a one-way sync flag, where the primary storage is treated as the only source of truth and the secondary storage is merely a mirror of the primary.

### What type of PR is it?
Improvement

### Todos

### What is the Jira issue?
<https://issues.apache.org/jira/browse/ZEPPELIN-1294>

### How should this be tested?
1. Setup two notebook storages through `ZEPPELIN_NOTEBOOK_STORAGE`, with the first is local file storage, i.e. `org.apache.zeppelin.notebook.repo.VFSNotebookRepo`. The second storage can be anything, e.g. `org.apache.zeppelin.notebook.repo.zeppelinhub.ZeppelinHubRepo`.
2. Add `export ZEPPELIN_NOTEBOOK_ONE_WAY_SYNC=true` to `zeppelin-env.sh`.
3. Start Zeppelin server.
4. Delete a notebook directly from the notebook folder.
5. Refresh notebooks from Zeppelin UI.
6. Observe from the secondary storage that the notebook has been deleted.

### Screenshots (if appropriate)

### Questions:
* Does the licenses files need update? No
* Is there breaking changes for older versions? No
* Does this needs documentation? Documentation for the new flag has been added to `install.md`

Author: Hao Xia <hao@optimizely.com>

Closes #1286 from jasonxh/hao/one-way-sync and squashes the following commits:

75a0670 [Hao Xia] Address PR comments
566b0ed [Hao Xia] Implement one-way sync for notebook repos